+++ src/L1/index.js
@@ -1,28 +1,24 @@
-export function add (a, b) {
-  return a + b
-}
-
-import * as utilsLevel1 from '.'
+import { add } from './utils/index'
 
 export function addNTimes (a, b, times) {
-  return utilsLevel1.add(a, b) * times
+  return add(a, b) * times
 }
 
 export function addWithCallback (a, b, callback) {
-  callback(utilsLevel1.add(a, b));
+  callback(add(a, b));
 }
 
 export function addAfter (a, b, time) {
   return new Promise((resolve) => {
     setTimeout(() => {
-      resolve(utilsLevel1.add(a, b));
+      resolve(add(a, b));
     }, time);  
   })
 }
 
 export function addAfterWithCallback (a, b, time, callback) {
   setTimeout(() => {
-    callback(utilsLevel1.add(a, b))
+    callback(add(a, b))
   }, time)
 }
