+++ src/L1/utils/index.js
@@ -1,0 +1,3 @@
+export function add (a, b) {
+  return a + b
+}
